Add files via upload
This is the Python code for SVM classifier using scikit-learn. It uses RBF, Polynomial and Linear kernels.
Model is chosen with highest estimated out of sample accuracy. That model is trained using full training data, predictions are made on the test data.
Ultimately, predictions are uploaded to the Kaggle for competition and I successfully got rank in top 28% in Kaggle competition in the group of 170 participants.